progress on ex6

+++ module_01/ex1/ft_garden_data.py
@@ -1,5 +1,5 @@
 class Plant:
-    def __init__(self, name: str, height: float, age: int) -> None:
+    def __init__(self, name: str, height: int, age: int) -> None:
         self.name = name
         self.height = height
         self.age = age
